donger
typecast

+++ GUI/Component/Wall.java
@@ -26,10 +26,10 @@ public class Wall extends GameObject {
 	@Override
 	public void paint(Graphics2D g) {
 		Stroke s = g.getStroke();
-		g.setStroke(new BasicStroke(WALL_WIDTH));
+		g.setStroke(new BasicStroke((float) WALL_WIDTH));
 		Point2D.Double loc = getLocation();
 		double dir = getDirection();
-		g.drawLine(loc.x, loc.y, loc.x+Game.CELL_SIDE*Math.cos(dir), loc.y+Game.CELL_SIDE*Math.sin(dir));
+		g.drawLine((int)loc.x,(int) loc.y,(int) (loc.x+Game.CELL_SIDE*Math.cos(dir)),(int) (loc.y+Game.CELL_SIDE*Math.sin(dir)));
 		g.setStroke(s);
 	}
